requirements.txt
requests
python-dateutil
Flask
click
reportlab
